+++ skoolkit/sna2skool.py
@@ -107,8 +107,10 @@ def main(args):
                        help='Specify the page (0-7) of a 128K snapshot to map to 49152-65535')
     group.add_argument('-t', '--text', dest='text', action='store_true',
                        help='Show ASCII text in the comment fields')
-    group.add_argument('-r', '--no-erefs', dest='write_refs', action='store_false',
+    group.add_argument('-r', '--no-erefs', dest='write_refs', action='store_const', const=-1, default=0,
                        help="Don't add comments that list entry point referrers")
+    group.add_argument('-R', '--erefs', dest='write_refs', action='store_const', const=1, default=0,
+                       help="Always add comments that list entry point referrers")
     group.add_argument('-n', '--defb-size', dest='defb_size', metavar='N', type=int, default=DEFB_SIZE,
                        help='Set the maximum number of bytes per DEFB statement to N (default={})'.format(DEFB_SIZE))
     group.add_argument('-m', '--defb-mod', dest='defb_mod', metavar='M', type=int, default=1,